LPD-24485 Inline JSONWebServiceRegistratorUtil into JSONWebServiceActionsManagerImpl
