+++ modules/apps/portal-remote/portal-remote-json-web-service-web/src/main/java/com/liferay/portal/remote/json/web/service/web/internal/JSONWebServiceActionsManagerImpl.java
@@ -9,27 +9,35 @@ import com.liferay.osgi.util.ServiceTrackerFactory;
 import com.liferay.petra.concurrent.DCLSingleton;
 import com.liferay.petra.lang.SafeCloseable;
 import com.liferay.petra.lang.ThreadContextClassLoaderUtil;
+import com.liferay.petra.reflect.AnnotationLocator;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.aop.AopService;
+import com.liferay.portal.kernel.bean.ClassLoaderBeanHandler;
+import com.liferay.portal.kernel.jsonwebservice.JSONWebService;
 import com.liferay.portal.kernel.jsonwebservice.JSONWebServiceAction;
 import com.liferay.portal.kernel.jsonwebservice.JSONWebServiceActionMapping;
 import com.liferay.portal.kernel.jsonwebservice.JSONWebServiceActionsManager;
+import com.liferay.portal.kernel.jsonwebservice.JSONWebServiceMode;
 import com.liferay.portal.kernel.jsonwebservice.NoSuchJSONWebServiceException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.service.ServiceWrapper;
 import com.liferay.portal.kernel.servlet.HttpMethods;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MethodParameter;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.ProxyUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.spring.aop.AopInvocationHandler;
 import com.liferay.portal.util.PropsValues;
 
+import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 
 import java.util.ArrayList;
@@ -64,6 +72,31 @@ import org.osgi.util.tracker.ServiceTrackerCustomizer;
 @Component(service = JSONWebServiceActionsManager.class)
 public class JSONWebServiceActionsManagerImpl
 	implements JSONWebServiceActionsManager {
+
+	public static void processBean(
+		JSONWebServiceActionsManager jsonWebServiceActionsManager,
+		String contextName, String contextPath, Object bean) {
+
+		if (!PropsValues.JSON_WEB_SERVICE_ENABLED) {
+			return;
+		}
+
+		JSONWebService jsonWebService = AnnotationLocator.locate(
+			_getTargetClass(bean), JSONWebService.class);
+
+		if (jsonWebService == null) {
+			return;
+		}
+
+		try {
+			_onJSONWebServiceBean(
+				jsonWebServiceActionsManager, contextName, contextPath, bean,
+				jsonWebService);
+		}
+		catch (Exception exception) {
+			_log.error(exception);
+		}
+	}
 
 	@Override
 	public Set<String> getContextNames() {
@@ -217,8 +250,7 @@ public class JSONWebServiceActionsManagerImpl
 
 		_ensureOpen();
 
-		JSONWebServiceRegistratorUtil.processBean(
-			this, contextName, contextPath, service);
+		processBean(this, contextName, contextPath, service);
 
 		int count = _getJSONWebServiceActionsCount(contextPath);
 
@@ -266,6 +298,105 @@ public class JSONWebServiceActionsManagerImpl
 	@Deactivate
 	protected void deactivate() {
 		_openedServiceTrackerDCLSingleton.destroy(ServiceTracker::close);
+	}
+
+	private static Class<?> _getTargetClass(Object service) {
+		while (ProxyUtil.isProxyClass(service.getClass())) {
+			InvocationHandler invocationHandler =
+				ProxyUtil.getInvocationHandler(service);
+
+			if (invocationHandler instanceof AopInvocationHandler) {
+				AopInvocationHandler aopInvocationHandler =
+					(AopInvocationHandler)invocationHandler;
+
+				service = aopInvocationHandler.getTarget();
+			}
+			else if (invocationHandler instanceof ClassLoaderBeanHandler) {
+				ClassLoaderBeanHandler classLoaderBeanHandler =
+					(ClassLoaderBeanHandler)invocationHandler;
+
+				Object bean = classLoaderBeanHandler.getBean();
+
+				if (bean instanceof ServiceWrapper) {
+					ServiceWrapper<?> serviceWrapper = (ServiceWrapper<?>)bean;
+
+					service = serviceWrapper.getWrappedService();
+				}
+				else {
+					service = bean;
+				}
+			}
+			else {
+				if (_log.isDebugEnabled()) {
+					_log.debug(
+						"Unable to handle proxy of type " + invocationHandler);
+				}
+
+				return null;
+			}
+		}
+
+		return service.getClass();
+	}
+
+	private static void _onJSONWebServiceBean(
+			JSONWebServiceActionsManager jsonWebServiceActionsManager,
+			String contextName, String contextPath, Object serviceBean,
+			JSONWebService jsonWebService)
+		throws Exception {
+
+		JSONWebServiceMode jsonWebServiceMode = JSONWebServiceMode.MANUAL;
+
+		if (jsonWebService != null) {
+			jsonWebServiceMode = jsonWebService.mode();
+		}
+
+		Method[] serviceMethods = JSONWebServiceScannerUtil.scan(serviceBean);
+
+		for (Method method : serviceMethods) {
+			JSONWebService methodJSONWebService = method.getAnnotation(
+				JSONWebService.class);
+
+			if (methodJSONWebService == null) {
+				if (!jsonWebServiceMode.equals(JSONWebServiceMode.AUTO)) {
+					continue;
+				}
+			}
+			else {
+				JSONWebServiceMode methodJSONWebServiceMode =
+					methodJSONWebService.mode();
+
+				if (methodJSONWebServiceMode.equals(
+						JSONWebServiceMode.IGNORE)) {
+
+					continue;
+				}
+			}
+
+			String httpMethod =
+				JSONWebServiceMappingResolverUtil.resolveHttpMethod(method);
+
+			if (!JSONWebServiceNamingUtil.isValidHttpMethod(httpMethod)) {
+				continue;
+			}
+
+			Class<?> serviceBeanClass = method.getDeclaringClass();
+
+			String path = JSONWebServiceMappingResolverUtil.resolvePath(
+				serviceBeanClass, method);
+
+			if (!JSONWebServiceNamingUtil.isIncludedPath(
+					contextName, contextPath, path)) {
+
+				continue;
+			}
+
+			if (JSONWebServiceNamingUtil.isIncludedMethod(method)) {
+				jsonWebServiceActionsManager.registerJSONWebServiceAction(
+					contextName, contextPath, serviceBean, serviceBeanClass,
+					method, path, httpMethod);
+			}
+		}
 	}
 
 	private boolean _addJSONWebServiceActionConfig(
